+++ examples/move_box_client/src/main.rs
@@ -103,13 +103,14 @@ fn on_connected(
     ui_state.log.push(format!("{name} connected"));
 
     game_state.set(GameState::Playing);
-    commands
-        .entity(entity)
-        .insert((SessionVisualizer::default(), TransportConfig {
+    commands.entity(entity).insert((
+        SessionVisualizer::default(),
+        TransportConfig {
             max_memory_usage: 64 * 1024,
             send_bytes_per_sec: 4 * 1024,
             ..default()
-        }));
+        },
+    ));
 }
 
 fn on_disconnected(