+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Made each crate have its own separate version
   - For now, all crates will share the same major and minor version number, but we are now free to bump the patch for individual subcrates
 - `aeronet_replicon` properly reports backend statistics to Replicon i.e. RTT, packet loss
+- Updated `bevy_replicon` to `0.30.0`
 
 # 0.11.0