+++ CHANGELOG.md
@@ -7,7 +7,9 @@
   - For now, all crates will share the same major and minor version number, but we are now free to bump the patch for individual subcrates
 - `aeronet_replicon` properly reports backend statistics to Replicon i.e. RTT, packet loss
 - Updated `bevy_replicon` to `0.30.0`
-- Update `wtransport` to `0.5.0`, `xwt-wtransport` and `xwt-web`
+- Updated `wtransport` to `0.5.0`, `xwt-wtransport` and `xwt-web`
+- Made integer/usize casts in `aeronet_transport` more safe and explicit
+  - Replace `LaneIndex(n)` with `LaneIndex::new(n)` (works in `const`)
 
 # 0.11.0